# src/routes/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "This is a sample server.  You can find me at\n[https://t.me/vinahradnicompot](https://t.me/vinahradnicompot)\n",
    "version": "1.0.0",
    "title": "Swagger"
  },
  "components": {
    "securitySchemes": {
      "Authorization": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/brands/{id}": {
      "get": {
        "tags": [
          "brands"
        ],
        "summary": "Get brand",
        "responses": {
          "200": {
            "description": "Ok"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "ID of the brand to get"
          }
        ]
      },
      "delete": {
        "tags": [
          "brands"
        ],
        "summary": "Delete brand",
        "responses": {
          "200": {
            "description": "Ok"
          },
          "401": {
            "description": "Invalid token"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "ID of the brand to delete"
          }
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      },
      "put": {
        "tags": [
          "brands"
        ],
        "summary": "Update brand",
        "responses": {
          "200": {
            "description": "Ok"
          },
          "401": {
            "description": "Invalid token"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "ID of the brand to update"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/brands": {
      "get": {
        "tags": [
          "brands"
        ],
        "summary": "Get brands",
        "responses": {
          "200": {
            "description": "Ok"
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "name",
            "schema": {
              "type": "string",
              "default": ""
            },
            "description": "The name of brand"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "number",
              "default": 1
            },
            "description": "The number of page"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "number",
              "default": 10
            },
            "description": "The number of items to limit"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "number",
              "default": 0
            },
            "description": "The number of items to skip"
          }
        ]
      },
      "post": {
        "tags": [
          "brands"
        ],
        "summary": "Add new brand",
        "responses": {
          "200": {
            "description": "Ok"
          },
          "401": {
            "description": "Invalid token"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "SwaggerHub API Auto Mocking",
      "url": "/"
    }
  ]
}